requirements.txt
absl-py==1.0.0
argon2-cffi==21.1.0
asgiref==3.5.0
astunparse==1.6.3
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
bokeh==2.4.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
crispy-bootstrap5==0.6
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
deepface==0.0.68
defusedxml==0.7.1
distlib==0.3.3
Django==4.0.1
django-allauth==0.47.0
django-bootstrap-v5==1.0.8
django-braces==1.15.0
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-star-ratings==0.9.2
djangorestframework==3.13.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.4.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.2
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mtcnn==0.1.1
munch==2.5.0
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pyproj==3.3.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==302
pywinpty==1.1.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.8
Send2Trash==1.8.0
setuptools-scm==6.3.2
Shapely==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0
